ai_collab_optimizer/requirements.txt
flask==2.3.3
pandas==2.0.0
networkx==3.1
openai==0.27.8
python-dotenv==1.0.0
numpy==1.24.3
python-dateutil==2.8.2
pytz==2023.3
tzdata==2023.3
